tts/requirements.txt
functions-framework==3.*
google-cloud-texttospeech==2.*